Separate event streams clients in different files
